LogStash/LogStash/requirements.txt
paho-mqtt
requests
pyyaml
python-logstash-async
